Reorganize latchkey wrapper and generalize permission flow
Permissions are eventually going to be a broader subsystem with multiple
backends; latchkey is just one of them. This commit consolidates the
latchkey-side code into a single class and generalizes the surface around
it so adding a second backend later is a refactor of internals only,
not a rename of every public name.

Latchkey wrapper consolidation:
- 'latchkey/gateway.py' is gone. 'latchkey/core.py' now holds a single
  Latchkey class that owns the gateway lifecycle (formerly
  LatchkeyGatewayManager) plus the services_info and auth_browser
  methods (formerly the LatchkeyServicesInfoProbe and
  LatchkeyAuthBrowserRunner classes from permission_flow.py).
- The LatchkeyGatewayDiscoveryHandler / DestructionHandler /
  ReconcileCallback move with it and lose the 'Gateway' infix; their
  field is now just 'latchkey'.
- 'latchkey/store.py' absorbs the contents of 'permissions_store.py'.
  The shared base error becomes LatchkeyStoreError.

Permission flow generalization:
- 'latchkey/permission_flow.py' moves to 'desktop_client/permissions.py'.
- GrantResult and GrantOutcome are gone; grant() now returns
  (was_granted, message). A failed sign-in is reported as a deny with a
  distinct message rather than a separate AUTH_FAILED status (which is
  also dropped from RequestStatus).
- PermissionGrantHandler takes a Latchkey directly instead of two
  separate probe/runner dependencies.

Renames driven by the same direction:
- templates/latchkey_permission.html -> templates/permissions.html
- render_latchkey_permission_dialog -> render_permission_dialog
- _handle_latchkey_permission_{grant,deny} -> _handle_permission_{grant,deny}
- app.state.latchkey_services_catalog  -> app.state.services_catalog
- app.state.latchkey_permission_handler -> app.state.permission_handler
- app.state.latchkey_gateway_manager   -> app.state.latchkey
- AgentCreator.latchkey_gateway_manager field -> AgentCreator.latchkey

Test layout follows the source layout: gateway_test renames to
core_test (with new services_info / auth_browser tests added back from
the deleted permission_flow_test), permissions_store_test merges into
store_test, permission_flow_test becomes permissions_test, and the
route integration test moves up to desktop_client/.
