+++ apps/minds/imbue/minds/desktop_client/agent_creator_test.py
@@ -19,8 +19,8 @@ from imbue.minds.desktop_client.agent_creator import clone_git_repo
 from imbue.minds.desktop_client.agent_creator import extract_repo_name
 from imbue.minds.desktop_client.agent_creator import make_log_callback
 from imbue.minds.desktop_client.agent_creator import run_mngr_create
-from imbue.minds.desktop_client.latchkey.gateway import AGENT_SIDE_LATCHKEY_PORT
-from imbue.minds.desktop_client.latchkey.gateway import LatchkeyGatewayManager
+from imbue.minds.desktop_client.latchkey.core import AGENT_SIDE_LATCHKEY_PORT
+from imbue.minds.desktop_client.latchkey.core import Latchkey
 from imbue.minds.desktop_client.latchkey.store import LatchkeyGatewayInfo
 from imbue.minds.errors import GitCloneError
 from imbue.minds.errors import GitOperationError
@@ -491,12 +491,12 @@ def test_agent_creator_cleans_up_pre_spawned_latchkey_gateway_on_failure(tmp_pat
     )
     fake_binary.chmod(0o755)
 
-    gateway_manager = LatchkeyGatewayManager(latchkey_binary=str(fake_binary))
-    gateway_manager.start(data_dir=tmp_path / "gateway-data")
+    latchkey = Latchkey(latchkey_binary=str(fake_binary))
+    latchkey.start(data_dir=tmp_path / "gateway-data")
     try:
         creator = AgentCreator(
             paths=WorkspacePaths(data_dir=tmp_path / "minds"),
-            latchkey_gateway_manager=gateway_manager,
+            latchkey=latchkey,
         )
         # "Local path" that does not exist -- mngr create will not even get
         # the chance to fail; _create_agent_background aborts with MngrCommandError.
@@ -513,6 +513,6 @@ def test_agent_creator_cleans_up_pre_spawned_latchkey_gateway_on_failure(tmp_pat
         creator.wait_for_all()
 
         # No lingering gateway or record for this agent.
-        assert gateway_manager.get_gateway_info(agent_id) is None
+        assert latchkey.get_gateway_info(agent_id) is None
     finally:
-        gateway_manager.stop()
+        latchkey.stop()
