+++ apps/minds/imbue/minds/desktop_client/runner.py
@@ -23,20 +23,18 @@ from imbue.minds.desktop_client.backend_resolver import MngrCliBackendResolver
 from imbue.minds.desktop_client.backend_resolver import MngrStreamManager
 from imbue.minds.desktop_client.cloudflare_client import CloudflareClient
 from imbue.minds.desktop_client.cloudflare_client import RemoteServiceConnectorUrl
-from imbue.minds.desktop_client.latchkey.gateway import LATCHKEY_BINARY
-from imbue.minds.desktop_client.latchkey.gateway import LatchkeyGatewayDestructionHandler
-from imbue.minds.desktop_client.latchkey.gateway import LatchkeyGatewayDiscoveryHandler
-from imbue.minds.desktop_client.latchkey.gateway import LatchkeyGatewayManager
-from imbue.minds.desktop_client.latchkey.gateway import LatchkeyGatewayReconcileCallback
-from imbue.minds.desktop_client.latchkey.permission_flow import LatchkeyAuthBrowserRunner
-from imbue.minds.desktop_client.latchkey.permission_flow import LatchkeyServicesInfoProbe
-from imbue.minds.desktop_client.latchkey.permission_flow import MngrMessageSender
-from imbue.minds.desktop_client.latchkey.permission_flow import PermissionGrantHandler
+from imbue.minds.desktop_client.latchkey.core import LATCHKEY_BINARY
+from imbue.minds.desktop_client.latchkey.core import Latchkey
+from imbue.minds.desktop_client.latchkey.core import LatchkeyDestructionHandler
+from imbue.minds.desktop_client.latchkey.core import LatchkeyDiscoveryHandler
+from imbue.minds.desktop_client.latchkey.core import LatchkeyReconcileCallback
 from imbue.minds.desktop_client.latchkey.services_catalog import LatchkeyServicesCatalogError
 from imbue.minds.desktop_client.latchkey.services_catalog import ServicePermissionInfo
 from imbue.minds.desktop_client.latchkey.services_catalog import load_services_catalog
 from imbue.minds.desktop_client.minds_config import MindsConfig
 from imbue.minds.desktop_client.notification import NotificationDispatcher
+from imbue.minds.desktop_client.permissions import MngrMessageSender
+from imbue.minds.desktop_client.permissions import PermissionGrantHandler
 from imbue.minds.desktop_client.request_events import RequestInbox
 from imbue.minds.desktop_client.request_events import load_response_events
 from imbue.minds.desktop_client.session_store import MultiAccountSessionStore
@@ -141,15 +139,14 @@ def start_desktop_client(
     backend_resolver = MngrCliBackendResolver()
     stream_manager = MngrStreamManager(resolver=backend_resolver)
     tunnel_manager = SSHTunnelManager()
-    latchkey_gateway_manager = _build_latchkey_gateway_manager(data_directory=data_directory)
-    latchkey_gateway_manager.start(data_dir=data_directory)
+    latchkey = _build_latchkey(data_directory=data_directory)
+    latchkey.start(data_dir=data_directory)
 
     minds_config = MindsConfig(data_dir=data_directory)
-    latchkey_services_catalog = _try_load_latchkey_services_catalog()
-    latchkey_permission_handler = PermissionGrantHandler(
+    services_catalog = _try_load_latchkey_services_catalog()
+    permission_handler = PermissionGrantHandler(
         data_dir=data_directory,
-        services_info_probe=LatchkeyServicesInfoProbe(latchkey_binary=LATCHKEY_BINARY),
-        auth_browser_runner=LatchkeyAuthBrowserRunner(latchkey_binary=LATCHKEY_BINARY),
+        latchkey=latchkey,
         mngr_message_sender=MngrMessageSender(),
     )
     cloudflare_client = _build_cloudflare_client(minds_config.remote_service_connector_url)
@@ -157,7 +154,7 @@ def start_desktop_client(
     agent_creator = AgentCreator(
         paths=paths,
         server_port=port,
-        latchkey_gateway_manager=latchkey_gateway_manager,
+        latchkey=latchkey,
     )
     telegram_orchestrator = TelegramSetupOrchestrator(paths=paths)
     is_electron = os.getenv("MINDS_ELECTRON") == "1"
@@ -210,19 +207,19 @@ def start_desktop_client(
     # subprocess for each discovered agent. For agents running in a container,
     # VM, or VPS the handler also sets up a reverse SSH tunnel so the agent
     # can reach the host-side gateway on a constant ``127.0.0.1`` URL.
-    latchkey_discovery_handler = LatchkeyGatewayDiscoveryHandler(
-        gateway_manager=latchkey_gateway_manager,
+    latchkey_discovery_handler = LatchkeyDiscoveryHandler(
+        latchkey=latchkey,
         tunnel_manager=tunnel_manager,
     )
-    latchkey_destruction_handler = LatchkeyGatewayDestructionHandler(gateway_manager=latchkey_gateway_manager)
+    latchkey_destruction_handler = LatchkeyDestructionHandler(latchkey=latchkey)
     stream_manager.add_on_agent_discovered_callback(latchkey_discovery_handler)
     stream_manager.add_on_agent_destroyed_callback(latchkey_destruction_handler)
 
     # Once the initial mngr-observe snapshot arrives, reconcile any adopted
     # gateways whose agent is no longer known so orphans from the previous
     # desktop-client session are cleaned up.
-    reconcile_callback = LatchkeyGatewayReconcileCallback(
-        gateway_manager=latchkey_gateway_manager,
+    reconcile_callback = LatchkeyReconcileCallback(
+        latchkey=latchkey,
         resolver=backend_resolver,
     )
     backend_resolver.add_on_change_callback(reconcile_callback)
@@ -237,7 +234,7 @@ def start_desktop_client(
         backend_resolver=backend_resolver,
         http_client=None,
         tunnel_manager=tunnel_manager,
-        latchkey_gateway_manager=latchkey_gateway_manager,
+        latchkey=latchkey,
         agent_creator=agent_creator,
         cloudflare_client=cloudflare_client,
         telegram_orchestrator=telegram_orchestrator,
@@ -248,8 +245,8 @@ def start_desktop_client(
         auth_backend_client=auth_backend_client,
         minds_config=minds_config,
         request_inbox=request_inbox,
-        latchkey_services_catalog=latchkey_services_catalog,
-        latchkey_permission_handler=latchkey_permission_handler,
+        services_catalog=services_catalog,
+        permission_handler=permission_handler,
         server_port=port,
         output_format=output_format,
     )
@@ -296,15 +293,15 @@ def _build_cloudflare_client(connector_url: AnyUrl) -> CloudflareClient:
     )
 
 
-def _build_latchkey_gateway_manager(data_directory: Path) -> LatchkeyGatewayManager:
-    """Build a ``LatchkeyGatewayManager`` honoring minds-level env var overrides.
+def _build_latchkey(data_directory: Path) -> Latchkey:
+    """Build a ``Latchkey`` wrapper honoring minds-level env var overrides.
 
     ``MINDS_LATCHKEY_BINARY`` can override the path to the ``latchkey`` CLI
     (typically supplied by the Electron shell, which installs the npm
     package under its own ``node_modules``). ``MINDS_LATCHKEY_DIRECTORY``
-    overrides the shared ``LATCHKEY_DIRECTORY`` that every spawned gateway
+    overrides the shared ``LATCHKEY_DIRECTORY`` that every spawned subprocess
     inherits; when unset we default to ``<minds_data_dir>/latchkey`` so all
-    gateways share a single credential store instead of scribbling into
+    invocations share a single credential store instead of scribbling into
     ``~/.latchkey``.
     """
     binary_override = os.environ.get("MINDS_LATCHKEY_BINARY")
@@ -316,7 +313,7 @@ def _build_latchkey_gateway_manager(data_directory: Path) -> LatchkeyGatewayMana
     else:
         latchkey_directory = data_directory / "latchkey"
 
-    return LatchkeyGatewayManager(
+    return Latchkey(
         latchkey_binary=latchkey_binary,
         latchkey_directory=latchkey_directory,
     )
