+++ apps/minds/imbue/minds/desktop_client/templates/permissions.html
